requirements.txt
openai==1.65.2
sentence-transformers==3.4.1
pandas==2.2.3
numpy==2.2.3
faiss-cpu==1.10.0
python-dotenv==1.0.1
streamlit==1.42.2
scikit-learn==1.6.1